// src/modules/auth/dto/verify-2fa.dto.ts
import { ApiProperty } from '@nestjs/swagger';
import { IsString, Length } from 'class-validator';

export class Verify2FADto {
  @ApiProperty({
    example: 'eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...',
    description: 'Temporary token issued after initial login, used to verify 2FA',
  })
  @IsString()
  tempToken: string;

  @ApiProperty({
    example: '123456',
    description: '6-digit code generated by the user’s 2FA app or sent via email/SMS',
  })
  @IsString()
  @Length(6, 6, { message: '2FA code must be exactly 6 digits' })
  code: string;
}